+++ src/util/dataTypes.js
@@ -40,9 +40,9 @@ export function normalize(value, isRoot = true) {
     value = value.trim()
   }
 
-  // Add spaces around operators inside calc() that do not follow an operator
+  // Add spaces around operators inside math functions like calc() that do not follow an operator
   // or '('.
-  value = value.replace(/calc\(.+\)/g, (match) => {
+  value = value.replace(/(calc|min|max|clamp)\(.+\)/g, (match) => {
     return match.replace(
       /(-?\d*\.?\d(?!\b-.+[,)](?![^+\-/*])\D)(?:%|[a-z]+)?|\))([+\-/*])/g,
       '$1 $2 '